# docs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "FastFuel",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5249"
    }
  ],
  "paths": {
...
    "/api/Order": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Gets all orders with optional filters.",
        "description": "**Required permission:** `Permission:Order:Read`",
        "operationId": "Order_GetAll",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "An optional order status filter.",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                }
              ]
            },
            "x-position": 1
          },
          {
            "name": "restaurantId",
            "in": "query",
            "description": "An optional restaurant ID filter.",
            "schema": {
              "type": "integer",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The matching orders, or a bad request response if a filter is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "x-required-permission": "Permission:Order:Read"
      },
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Makes a new entity.",
        "description": "**Required permission:** `Permission:Order:Create`",
        "operationId": "Order_Create",
        "requestBody": {
          "x-name": "requestDto",
          "description": "The entity data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "The created entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "x-required-permission": "Permission:Order:Create"
      }
    },
    "/api/Order/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Gets one entity.",
        "description": "**Required permission:** `Permission:Order:Read`",
        "operationId": "Order_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The entity id.",
            "schema": {
              "type": "integer"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The entity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "x-required-permission": "Permission:Order:Read"
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Deletes one entity.",
        "description": "**Required permission:** `Permission:Order:Delete`",
        "operationId": "Order_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The entity id.",
            "schema": {
              "type": "integer"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "x-required-permission": "Permission:Order:Delete"
      }
    },
    "/api/Order/my": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Gets the orders of the currently authenticated user.",
        "operationId": "Order_GetMyOrders",
        "responses": {
          "200": {
            "description": "The current user's orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/Order/{id}/status": {
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "Updates the status of an existing order.",
        "description": "**Required permission:** `Permission:Order:UpdateStatus`",
        "operationId": "Order_UpdateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The identifier of the order to update.",
            "schema": {
              "type": "integer"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "orderStatus",
          "description": "The new order status.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": "No content when the status update succeeds; otherwise an error response."
          },
          "404": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "x-required-permission": "Permission:Order:UpdateStatus"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OrderStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Pending",
          "InProgress",
          "Ready",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          "Pending",
          "InProgress",
          "Ready",
          "Completed",
          "Cancelled"
        ]
      },
...
      "OrderResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "restaurantId",
          "orderNumber",
          "status",
          "createdAt",
          "menus",
          "foods",
          "id"
        ],
        "properties": {
          "userId": {
            "type": "integer",
            "description": "The user identifier.\n            "
          },
          "restaurantId": {
            "type": "integer",
            "description": "The restaurant identifier.\n            "
          },
          "orderNumber": {
            "type": "integer",
            "description": "The public order number.\n            "
          },
          "status": {
            "description": "The status of the order.\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/OrderStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "The date and time when the item was created.\n            ",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "description": "When the order is completed at.\n            ",
            "format": "date-time",
            "nullable": true
          },
          "menus": {
            "type": "array",
            "description": "The list of menus.\n            ",
            "items": {
              "$ref": "#/components/schemas/OrderMenuResponseDto"
            }
          },
          "foods": {
            "type": "array",
            "description": "The list of foods.\n            ",
            "items": {
              "$ref": "#/components/schemas/OrderFoodResponseDto"
            }
          },
          "id": {
            "type": "integer",
            "description": "The unique identifier.\n            "
          }
        }
      },
      "OrderMenuResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "menuId",
          "originalMenuName",
          "originalMenuPrice",
          "quantity"
        ],
        "properties": {
          "menuId": {
            "type": "integer",
            "description": "The menu identifier.\n            "
          },
          "originalMenuName": {
            "type": "string",
            "description": "The original menu name at the time of order placement.\n            "
          },
          "originalMenuPrice": {
            "type": "number",
            "description": "The original menu price at the time of order placement.\n            ",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "description": "The quantity of the menu in the order.\n            "
          },
          "specialInstructions": {
            "type": "string",
            "description": "Additional instructions for preparation.\n            ",
            "nullable": true
          }
        }
      },
      "OrderFoodResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "foodId",
          "originalFoodName",
          "originalFoodPrice",
          "quantity"
        ],
        "properties": {
          "foodId": {
            "type": "integer",
            "description": "The food identifier.\n            "
          },
          "originalFoodName": {
            "type": "string",
            "description": "The original food name at the time of order placement.\n            "
          },
          "originalFoodPrice": {
            "type": "number",
            "description": "The original food price at the time of order placement.\n            ",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "description": "The quantity of the food in the order.\n            "
          },
          "specialInstructions": {
            "type": "string",
            "description": "Additional instructions for preparation.\n            ",
            "nullable": true
          }
        }
      },
      "OrderRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "restaurantId",
          "menus",
          "foods"
        ],
        "properties": {
          "restaurantId": {
            "type": "integer",
            "description": "The restaurant identifier.\n            "
          },
          "menus": {
            "type": "array",
            "description": "The list of menus.\n            ",
            "items": {
              "$ref": "#/components/schemas/OrderMenuRequestDto"
            }
          },
          "foods": {
            "type": "array",
            "description": "The list of foods.\n            ",
            "items": {
              "$ref": "#/components/schemas/OrderFoodRequestDto"
            }
          }
        }
      },
      "OrderMenuRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "menuId",
          "quantity"
        ],
        "properties": {
          "menuId": {
            "type": "integer",
            "description": "The menu identifier.\n            "
          },
          "quantity": {
            "type": "integer",
            "description": "The quantity of the menu in the order.\n            "
          },
          "specialInstructions": {
            "type": "string",
            "description": "Additional instructions for preparation.\n            ",
            "nullable": true
          }
        }
      },
      "OrderFoodRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "foodId",
          "quantity"
        ],
        "properties": {
          "foodId": {
            "type": "integer",
            "description": "The food identifier.\n            "
          },
          "quantity": {
            "type": "integer",
            "description": "The quantity of the food in the order.\n            "
          },
          "specialInstructions": {
            "type": "string",
            "description": "Additional instructions for preparation.\n            ",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}